refactor: rename options and move http middleware
