+++ options/doc.go
@@ -1,2 +1,2 @@
 // Package option provides a set of options for the client.
-package option
+package options
